Move timeout out of managed_dict, and serialise a dict instead of a tuple.
Move timeout out of managed_dict, as it is metadata that is not part of
the session. Previously any timeout set via
.adjust_timeout_for_session() would be lost after .clear(), for example.

Serialise session dict, created time, and timeout for storage in Redis
as a dict instead of a tuple, as a dict is more extensible, more
readable and less bug-prone.