Fix invalidate.
This fix adds support for the use of a new session within the same
request after an .invalidate() call. invalidate() now deletes the
session_id key from Redis, and RedisSession now knows to start a new
session with a new session_id when it is needed. (Previously
invalidate() only cleared the session dict and set a header to delete
the cookie, and did not take into account the use of the session after
invalidate().)

It also fixes the cookie setting after invalidate(). Previously
invalidate() just added another Set-Cookie header to the response to
delete the cookie, meaning there could be more than one Set-Cookie
headers for the same cookie name, which should not happen according to
RFC 6265. This has been fixed to set one and only one Set-Cookie header
(if any is needed) in the response callback, and to handle the various
combinations of whether the request started with a new session or
existing session, whether a new session was used after invalidate(),
more than one invalidate()s in the same request, and the
cookie_on_exception setting.