Change asserts to use a different approach.
To assert that a header is one that sets a cookie as opposed to one that
deletes a cookie, we no longer assert the serialized session_id. This
means the tests would not break if `response.set_cookie()` makes any
changes to the serialized session_id before inserting it into the
cookie, as WebOb<=1.3 did by quoting certain characters.

The alternative approach chosen is to assert that 'Max-Age=0' is not in
the header string. A Set-Cookie header is either one to set cookie (in
which case 'Max-Age=0' would not be in the string), or one to delete
cookie (in which case 'Max-Age=0' would be in the string.)

It is a negative assertion, but all the alternatives would make the
tests too complicated and/or difficult to make sense of. This is the
least bad approach I can see for now.